build: Add premake4 build files and rtc gitignores
Premake files were generated by running the gdal configure
script.  The script uses detection in order to support as much
as possible on a native environment.  For the runtimecore,
we only support a small subset of functionality. You can use
options in the configure script to limit some of this functionality
but a lot of it cannot be toggled so you need to undefine a lot
of the GDAL FRMT defines and not build them. If a format is
requested in the future, the defines can be added and files can
start being added as well as determined by the link line.