Add gdal auto-generated version and config files needed to build
GDAL has a few autogenerated files that are produced when the configure
script is run. The configure script was run with:

./configure --with-cpp14 --without-libtool --without-ld-shared \
--with-pic --disable-shared --with-libz=external \
--with-png=${path_to_png} --with-libtiff=${path_to_libtiff} \
--with-jpeg=${path_to_jpeg} --with-gif=${path_to_giflib} \
--with-expat-inc=${path_to_expat} --without-pcraster \
--without-crypto --without-grib --without-mrf --without-pcidsk \
--without-pcre --without-qhull

When cross compiling, you need to set CXX, CXXFLAGS
and CC while changing the build target. For example, for android,
run ./build.sh on android with -v in order to output the compiler
and our common flags.  Once these are set, run configure with
the additional arguments for host and build. So on linux cross
compiling for armv7, you can add
--host=x86_64-pc-linux-gnu --build=armv8